chaining 50%; about to do mocks

+++ src/main/java/util/exceptions/ExceptionDemo.java
@@ -1,8 +1,8 @@
 package util.exceptions;
 
 public class ExceptionDemo {
-    public static double generalException(int number, int divisor) {
-        return (float) number / divisor;
+    public static int generalException(int number, int divisor) {
+        return number / divisor;
     }
     public static double userException(int number, int divisor) throws DivideByZeroException {
         if(divisor == 0) throw new DivideByZeroException("Divided " + number + " by " + divisor);
@@ -15,3 +15,5 @@ class DivideByZeroException extends Exception {
         super(errorMessage);
     }
 }
+
+
